Update FrontAndRearLights.ino
I was having an issue with the included Arduino sketch for the Front and Rear LED Strip lights. The lights were showing GREEN, RED, BLUE, and OFF in that order and it would repeat across the strip.

I discovered that the strip in the build of materials does not contain a white LED and the sketch was calling for a strip that contained GRBW. I changed it to GRB only and adjusted the front LED Color to (255,255,255) to achieve the white color. The new sketch I used is forked here.